Implementation of text renderer
