+++ src/lib_ccx/ccx_encoders_curl.c
@@ -1,8 +1,6 @@
 #ifdef WITH_LIBCURL
 #include "ccx_decoders_common.h"
 #include "ccx_encoders_common.h"
-#include "spupng_encoder.h"
-#include "ccx_encoders_spupng.h"
 #include "utility.h"
 #include "ocr.h"
 #include "ccx_decoders_608.h"
