+++ src/lib_ccx/ccx_encoders_smptett.c
@@ -25,7 +25,6 @@
 #include "ccx_common_option.h"
 #include "ccx_encoders_common.h"
 #include "png.h"
-#include "spupng_encoder.h"
 #include "ocr.h"
 #include "utility.h"
 #include "ccx_encoders_helpers.h"
